requirements.txt
streamlit
torch
torchvision
numpy
timm
Pillow
matplotlib
requests
os